fix(pwa): keep the node:fs blog loader out of the client bundle
The blog page components imported formatBlogDate from lib/blog.ts, which
also statically imports node:fs/promises. That pulled the loader into the
browser bundle and Turbopack failed the prod build ("does not support
external modules (request: node:fs/promises)").

Split the client-safe types + formatBlogDate into lib/blogTypes.ts (no
node: imports). lib/blog.ts stays server-only and is now referenced only
from getStaticProps/getStaticPaths, which Next strips from the client
bundle — matching the working /guides pipeline.
