+++ CLAUDE.md
@@ -76,7 +76,7 @@ Aura/                 # repo/folder name is unchanged; the product is "Madori"
 - **Web Push frontend (#100)**: Completes the push pipeline begun by the backend `PushSubscription` / `App\Push\WebPushSender`. `pwa/public/sw.js` (service worker, registered from `_app.tsx`) renders a `PushPayload` (`{title, body, url, tag}`) as a desktop notification and focuses/opens the target URL on click. `pwa/lib/push.ts` `enablePush()` runs the opt-in: request permission → `PushManager.subscribe()` with the VAPID public key → `POST /me/push-subscriptions`. `pushNotificationsEnabled` defaults to **true**; `EnablePushPrompt` (`pwa/components/notifications/EnablePushPrompt.tsx`) shows a small amber warning in the `RemindersEditor` when a task has reminders but the browser isn't granted, with an inline "Turn on" — its "Not now" never disables the account-level setting (public-computer safe). **VAPID keys**: the public key is committed in `pwa/.env.production` as `NEXT_PUBLIC_VAPID_PUBLIC_KEY` (shipped to browsers — not secret; baked into the PWA at CI build); the API reads `VAPID_PUBLIC_KEY` / `VAPID_PRIVATE_KEY` / `VAPID_SUBJECT` via `%env(default::VAPID_*)%` (`WebPushSender` no-ops with a warning when blank). In prod these live in the server's untracked `/opt/aura/.env`, wired to `php` + `worker` via `compose.yaml` `${VAPID_*}` refs (the worker sends reminder/digest push, so it needs them too).
 - **Activity timeline (PWA)**: Activity streams (`GET /tasks|projects|spaces/{id}/activity`) render through `pwa/components/activity/ActivityTimeline.tsx` — a responsive "history point graph": a vertical rail of action-coloured markers (create=emerald, update=sky, remove=red) with one event card each, using CSS container queries (`@container`) so cards alternate to both sides when wide (≥42rem) and stack on the right when narrow (so it fits the task drawer and a full-width tab alike). `ActivityPanel` feeds it and formats change values (PHP `DateTime` → localised, bare `{id}` references dropped, booleans → yes/no). Activity rolls **up** the hierarchy: `App\Service\ProjectActivityScope` resolves a project's groups (project + tasks + custom-field definitions, incl. deleted children recovered via the versioned `project` FK) and is shared by `ProjectActivityController` and `SpaceActivityController` (which unions it across every project), so a space sees everything its projects do — like space membership cascades.
 - **Project Settings tab (PWA)**: The project detail page's Settings tab is sub-sectioned with a left menu (mirroring the user settings page): **Privacy** (space + members), **Custom fields** (the `CustomFieldsManager`), **Danger zone** (move/copy + delete project). The custom-field change log is reached via a history icon on the fields list (the standalone `/projects/{id}/custom-fields` page was removed).
-- **Blog (file-based, PWA-only)**: Public blog (tips/tricks/progress updates) built as an SEO surface — **no backend, nothing in the DB**. Posts are markdown files committed under `pwa/content/blog/<slug>.md` with frontmatter (`title`, optional `slug` defaulting to filename, `description` = SEO meta, `date`, optional `author`, `draft` bool, optional `ogImage` path under `pwa/public/blog/`), so content is versioned/diffable/PR-reviewed. Chosen over a DB CMS for version control; the trade-off is publish=deploy (content is static, baked at build). Everything resolves at **build time** — the prod image ships only `public/` + `.next`, never `content/`, so nothing reads markdown at request time. `pwa/lib/blog.ts` is the server-only loader (`getAllPosts`/`getPostSummaries`/`getPostBySlug` reading `content/blog`, a tiny dependency-free frontmatter parser, plus `siteUrl()` + `formatBlogDate()`) imported only from `getStaticProps`/`getStaticPaths`. `pwa/pages/blog/index.tsx` (SSG listing) + `pwa/pages/blog/[slug].tsx` (SSG detail, `getStaticPaths` `fallback:false`, full SEO `<head>`: description, canonical, OpenGraph, Twitter card, `article:published_time`). Drafts (`draft:true`) render in `pnpm dev` for preview but are excluded from prod builds/listing/paths/sitemap. `pwa/scripts/gen-blog-sitemap.mjs` is a predev/prebuild hook (chained after `sync-docs` in `pwa/package.json`) that writes the git-ignored `pwa/public/sitemap.xml` from frontmatter; `pwa/public/robots.txt` advertises it. Absolute canonical/OG/sitemap URLs come from `NEXT_PUBLIC_SITE_URL` (committed in `pwa/.env.production`, default `http://localhost:3000`). Footer links to `/blog`. See `docs/developer/blog.md`.
+- **Blog (file-based, PWA-only)**: Public blog (tips/tricks/progress updates) built as an SEO surface — **no backend, nothing in the DB**. Posts are markdown files committed under `pwa/content/blog/<slug>.md` with frontmatter (`title`, optional `slug` defaulting to filename, `description` = SEO meta, `date`, optional `author`, `draft` bool, optional `ogImage` path under `pwa/public/blog/`), so content is versioned/diffable/PR-reviewed. Chosen over a DB CMS for version control; the trade-off is publish=deploy (content is static, baked at build). Everything resolves at **build time** — the prod image ships only `public/` + `.next`, never `content/`, so nothing reads markdown at request time. `pwa/lib/blog.ts` is the **server-only** loader (`getAllPosts`/`getPostSummaries`/`getPostBySlug`/`siteUrl` reading `content/blog`, a tiny dependency-free frontmatter parser); it imports `node:fs` so it must be referenced ONLY from `getStaticProps`/`getStaticPaths` — using it in a component pulls `fs` into the client bundle and fails the Turbopack build. Client-safe types + `formatBlogDate()` live in `pwa/lib/blogTypes.ts` (no `node:` imports) for the page components. `pwa/pages/blog/index.tsx` (SSG listing) + `pwa/pages/blog/[slug].tsx` (SSG detail, `getStaticPaths` `fallback:false`, full SEO `<head>`: description, canonical, OpenGraph, Twitter card, `article:published_time`). Drafts (`draft:true`) render in `pnpm dev` for preview but are excluded from prod builds/listing/paths/sitemap. `pwa/scripts/gen-blog-sitemap.mjs` is a predev/prebuild hook (chained after `sync-docs` in `pwa/package.json`) that writes the git-ignored `pwa/public/sitemap.xml` from frontmatter; `pwa/public/robots.txt` advertises it. Absolute canonical/OG/sitemap URLs come from `NEXT_PUBLIC_SITE_URL` (committed in `pwa/.env.production`, default `http://localhost:3000`). Footer links to `/blog`. See `docs/developer/blog.md`.
 - **Key packages**: doctrine/orm, nelmio/cors-bundle, symfony/security-bundle, symfony/serializer, symfony/messenger
 
 ### PWA (Frontend)
